moved "GitClass3.java" file into a folder
