--- src/File1/sima_file/GitClass2.java
@@ -1,9 +1,0 @@
-package File1.sima_file;
-
-public class GitClass2 {
-
-    public static void main(String[] args) {
-
-        System.out.println("This is a class for practice");
-    }
-}
